updated french, changed on/off duty language, fixed 2 bugs
Changing to new language: on duty/off duty to at sea/not at sea
Found a bug where "search fishery" wasn't used, because it should be "search fisheries", and another issue where a period needed to be added to match the string in the code.

Updated French: on duty/at sea but also the "Search EMS" string was too long.